+++ demo/src/main/java/com/nabla/sdk/demo/scene/MainActivity.kt
@@ -1,46 +1,15 @@
 package com.nabla.sdk.demo.scene
 
-import android.content.Intent
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.lifecycleScope
 import com.nabla.sdk.demo.databinding.ActivityMainBinding
-import com.nabla.sdk.demo.scene.ConversationActivity.Companion.CONVERSATION_ID_EXTRA
-import com.nabla.sdk.messaging.core.NablaMessagingClient
-import com.nabla.sdk.messaging.ui.helper.ConversationListViewModelFactory
-import com.nabla.sdk.messaging.ui.scene.conversations.ConversationListViewModel
-import com.nabla.sdk.messaging.ui.scene.conversations.bindViewModel
-import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
-
-    private val viewModel: ConversationListViewModel by viewModels {
-        ConversationListViewModelFactory(owner = this)
-    }
-
     private lateinit var binding: ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-        binding.createConversation.setOnClickListener {
-            lifecycleScope.launch {
-                NablaMessagingClient.getInstance().createConversation()
-                    .getOrNull()
-            }
-        }
-
-        binding.conversationListView.bindViewModel(
-            viewModel,
-            onConversationClicked = { id ->
-                startActivity(
-                    Intent(applicationContext, ConversationActivity::class.java)
-                        .apply { putExtra(CONVERSATION_ID_EXTRA, id.value) }
-                )
-            }
-        )
     }
 }