+++ messaging-ui-tests/src/main/java/com/nabla/sdk/uitests/scene/MainActivity.kt
@@ -2,7 +2,6 @@ package com.nabla.sdk.uitests.scene
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
 import com.nabla.sdk.uitests.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
@@ -16,16 +15,8 @@ class MainActivity : AppCompatActivity() {
         if (savedInstanceState == null) {
             supportFragmentManager
                 .beginTransaction()
-                .replace(binding.fragmentContainer.id, ConversationsFragment())
+                .replace(binding.fragmentContainer.id, StubbedInboxFragment())
                 .commit()
         }
-    }
-
-    fun pushFragment(fragment: Fragment) {
-        supportFragmentManager
-            .beginTransaction()
-            .replace(binding.fragmentContainer.id, fragment)
-            .addToBackStack(null)
-            .commit()
     }
 }